Modified mar to use unsigned instead of size_t types.
git-svn-id: https://aug.svn.sourceforge.net/svnroot/aug/trunk@88 a3a97f7c-8e16-0410-9893-1d5f7d0f8f6e

+++ src/c/augmar/body_.h
@@ -15,20 +15,20 @@ struct aug_info_;
 
 AUGMAR_EXTERN int
 aug_setcontent_(aug_seq_t seq, struct aug_info_* info, const void* data,
-                size_t size);
+                unsigned size);
 
 AUGMAR_EXTERN int
-aug_truncate_(aug_seq_t seq, struct aug_info_* info, size_t size);
+aug_truncate_(aug_seq_t seq, struct aug_info_* info, unsigned size);
 
-AUGMAR_EXTERN ssize_t
-aug_write_(aug_seq_t seq, struct aug_info_* info, size_t offset,
-           const void* buf, size_t size);
+AUGMAR_EXTERN int
+aug_write_(aug_seq_t seq, struct aug_info_* info, unsigned offset,
+           const void* buf, unsigned size);
 
 AUGMAR_EXTERN const void*
 aug_content_(aug_seq_t seq, const struct aug_info_* info);
 
-AUGMAR_EXTERN ssize_t
-aug_read_(aug_seq_t seq, const struct aug_info_* info, size_t offset,
-          void* buf, size_t size);
+AUGMAR_EXTERN int
+aug_read_(aug_seq_t seq, const struct aug_info_* info, unsigned offset,
+          void* buf, unsigned size);
 
 #endif /* AUGMAR_BODY_H_ */
