+++ src/c/augmar/header_.h
@@ -18,37 +18,37 @@ aug_removefields_(aug_seq_t seq, struct aug_info_* info);
 
 AUGMAR_EXTERN int
 aug_setfield_(aug_seq_t seq, struct aug_info_* info,
-              const struct aug_field* field, size_t* ord);
+              const struct aug_field* field, unsigned* ord);
 
 AUGMAR_EXTERN int
-aug_setvalue_(aug_seq_t seq, struct aug_info_* info, size_t ord,
-              const void* value, size_t size);
+aug_setvalue_(aug_seq_t seq, struct aug_info_* info, unsigned ord,
+              const void* value, unsigned size);
 
 AUGMAR_EXTERN int
 aug_unsetbyname_(aug_seq_t seq, struct aug_info_* info, const char* name,
-                 size_t* ord);
+                 unsigned* ord);
 
 AUGMAR_EXTERN int
-aug_unsetbyord_(aug_seq_t seq, struct aug_info_* info, size_t ord);
+aug_unsetbyord_(aug_seq_t seq, struct aug_info_* info, unsigned ord);
 
 AUGMAR_EXTERN const void*
 aug_valuebyname_(aug_seq_t seq, const struct aug_info_* info,
-                 const char* name, size_t* size);
+                 const char* name, unsigned* size);
 
 AUGMAR_EXTERN const void*
-aug_valuebyord_(aug_seq_t seq, const struct aug_info_* info, size_t ord,
-                size_t* size);
+aug_valuebyord_(aug_seq_t seq, const struct aug_info_* info, unsigned ord,
+                unsigned* size);
 
 AUGMAR_EXTERN int
 aug_field_(aug_seq_t seq, const struct aug_info_* info,
-           struct aug_field* field, size_t ord);
+           struct aug_field* field, unsigned ord);
 
 AUGMAR_EXTERN int
 aug_ordtoname_(aug_seq_t seq, const struct aug_info_* info, const char** name,
-               size_t ord);
+               unsigned ord);
 
 AUGMAR_EXTERN int
-aug_nametoord_(aug_seq_t seq, const struct aug_info_* info, size_t* ord,
+aug_nametoord_(aug_seq_t seq, const struct aug_info_* info, unsigned* ord,
                const char* name);
 
 #endif /* AUGMAR_HEADER_H_ */
